backend/pyproject.toml
[project]
name = "tailormade-coloring-book"
version = "1.0.0"
description = "TailorMade Coloring Book — AI-powered personalized coloring book generator"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pillow>=11.0.0",
    "fal-client>=0.5.0",
    "firebase-admin>=6.6.0",
    "anthropic>=0.40.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "weasyprint>=62.0",
    "boto3>=1.35.0",
    "aiofiles>=24.0.0",
    "httpx>=0.27.0",
    "tenacity>=9.1.4",
    "cachetools>=7.0.1",
    "sentry-sdk[fastapi]>=2.52.0",
    "redis>=7.1.1",
    "celery>=5.6.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
